requirements.txt
# --- Core functionality ---
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==11.0.0
numpy==2.1.3
tqdm==4.67.1

# --- CLI & packaging ---
click==8.1.7
pyinstaller==6.11.0
nuitka==2.5.5

# --- Development tools ---
ruff==0.9.4
zstandard==0.24.0